+++ src/Simpletools/Db/Cassandra/SessionHandler.php
@@ -9,7 +9,7 @@ class SessionHandler implements \SessionHandlerInterface, \SessionIdInterface
     protected $_sessionTable;
 
     //time in sec, defaults to php ini session.gc_maxlifetime, however if that can't be read from php.ini, it will fail-over to 1800 sec
-    protected $_maxLifeTime = 1800;
+    protected static $_maxLifeTime = 1800;
 
     protected static $_settings;
 
@@ -17,31 +17,83 @@ class SessionHandler implements \SessionHandlerInterface, \SessionIdInterface
     {
         if(isset($settings['maxLifeTime']))
         {
-            $this->_maxLifeTime = $settings['maxLifeTime'];
+            self::$_maxLifeTime = $settings['maxLifeTime'];
         }
         elseif(isset(self::$_settings['maxLifeTime']) && self::$_settings['maxLifeTime'])
         {
-            $this->_maxLifeTime = self::$_settings['maxLifeTime'];
+            self::$_maxLifeTime = self::$_settings['maxLifeTime'];
         }
         elseif (get_cfg_var("session.gc_maxlifetime"))
         {
-            $this->_maxLifeTime = get_cfg_var("session.gc_maxlifetime");
+            self::$_maxLifeTime = get_cfg_var("session.gc_maxlifetime");
         }
 
-        $this->_maxLifeTime = (int) $this->_maxLifeTime;
+        self::$_maxLifeTime = (int) self::$_maxLifeTime;
 
-        if($this->_maxLifeTime<1)
+        if(self::$_maxLifeTime<1)
             throw new Exception('maxLifeTime can\'t be smaller than 1 due to TimeWindowCompactionStrategy');
     }
 
     public function maxLifeTime($seconds=null)
     {
         if(!$seconds)
-            return $this->_maxLifeTime;
+            return self::$_maxLifeTime;
         else
-            $this->_maxLifeTime = $seconds;
+            self::$_maxLifeTime = $seconds;
 
         return $this;
+    }
+
+    public static function regenerateSessionId()
+    {
+        if(session_status()!=PHP_SESSION_ACTIVE)
+        {
+            session_start();
+        }
+
+        $currentId = session_id();
+
+        $newId = (string) new Uuid();
+
+        try {
+            $currentSession = (new Doc($currentId))
+                ->client(new Client(self::$_settings['cluster']))
+                ->keyspace(self::$_settings['keyspace'])
+                ->table(self::$_settings['table'])
+                ->load();
+        }
+        catch(\Exception $e)
+        {
+            return false;
+        }
+
+        $batch = (new Batch())
+            ->client(new Client(self::$_settings['cluster']));
+
+        if(@self::$_settings['consistency'])
+            $batch->consistency(self::$_settings['consistency']);
+
+        if(@self::$_settings['writeConsistency'])
+            $batch->consistency(self::$_settings['writeConsistency']);
+
+        //session_unset();
+        //session_destroy();
+        session_write_close();
+        //setcookie(session_name(),'',0,'/');
+
+        $batch
+            ->add(
+                (new Query(self::$_settings['table'],self::$_settings['keyspace']))->delete('id',$currentId)
+            )
+            ->add(
+                (new Query(self::$_settings['table'],self::$_settings['keyspace']))->set([
+                    'id'                =>  $newId,
+                    'data'              =>  (string) $currentSession->body->data
+                ])->expires(self::$_maxLifeTime))
+        ->run();
+
+        session_id($newId);
+        session_start();
     }
 
     public static function settings($settings)
@@ -139,7 +191,7 @@ class SessionHandler implements \SessionHandlerInterface, \SessionIdInterface
         $q->set([
             'id'                =>  $id,
             'data'              =>  $data
-        ])->expires($this->_maxLifeTime)->run();
+        ])->expires(self::$_maxLifeTime)->run();
 
         unset($q);